pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grandmofty-mcp"
version = "0.1.0"
description = "MCP server for querying Metal-Organic Framework databases"
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
authors = [{name = "Reto Stamm", email = "reto@retostamm.com"}]
keywords = ["mcp", "mof", "chemistry", "materials-science", "database"]
dependencies = [
    "chemdb-common>=0.1.0",
    "mcp>=1.0",
    "mofdb-client>=0.9",
]

[project.optional-dependencies]
postgres = [
    "psycopg[binary]>=3.0",
]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.5",
]

[project.scripts]
grandmofty-mcp = "grandmofty.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/grandmofty"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true
tag_name = "v{new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/grandmofty/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'
